forceFetch the7 tiau7
