+++ src/Iah/Kong/其他建議.jsx
@@ -50,8 +50,10 @@ export default class 其他建議 extends React.Component {
     let 文本 = 其他建議.map((建議, i) => (
         <一个建議 key={i}
           後端網址={this.props.後端網址}
-          文本資料={建議.文本資料} 音標資料={建議.音標資料}
-          來開例句={this.props.開例句.bind(this, this.props.華語關鍵字, 建議.文本資料, 建議.音標資料)}/>
+          文本資料={建議.文本資料}
+          音標資料={建議.音標資料}
+          來開例句={this.props.開例句.bind(this, this.props.華語關鍵字, 建議.文本資料, 建議.音標資料)}
+          variables={建議} />
     ));
     if (文本.length == 0) {
       return null;
